requirements.txt
Django==3.0.4
psycopg2==2.8.4
# uwsgi==2.0.18
django-bootstrap4==21.1
django-summernote==0.8.20.0
django-import-export==2.7.0
django-cleanup==5.2.0
django-imagekit==4.1.0
Pillow==8.4.0
psycopg2-binary
gunicorn